backend/requirements.txt
# Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Data validation
pydantic>=2.7.1
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.20.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Agent orchestration
langgraph>=0.2.0
langchain-groq>=0.1.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0

# ML and data processing
interpret>=0.6.0
scikit-learn>=1.5.0
pandas>=2.2.3
numpy>=2.1.0

# OCR and document processing
pytesseract>=0.3.10
Pillow>=10.4.0
opencv-python>=4.10.0.84
pdf2image>=1.17.0
PyPDF2>=3.0.1

# RAG / vector store
faiss-cpu>=1.8.0
pypdf>=3.17.0

# Utilities
tenacity>=8.2.3
Faker>=24.0.0
slowapi>=0.1.9

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
